refactored the createGroup endpoint
