+++ apps/server/src/controllers/groupControllers/createGroup.ts
@@ -12,126 +12,11 @@ import { transactionType } from '../../models/transactions'
 import { createGroupSchema } from '../../utils/validators'
 import { v2 as cloudinary } from 'cloudinary'
 import { createNewMember } from './helpers'
-import nodemailer from 'nodemailer'
-import { addMilliseconds, format } from 'date-fns'
-import User from '../../models/users'
-
-async function notifyBeneficiary(
-  userId: string | null,
-  user: User
-): Promise<void> {
-  if (userId) {
-    console.log(`Notifying beneficiary with userId: ${userId}`)
-
-    // Create a nodemailer transporter
-    const transporter = nodemailer.createTransport({
-      service: 'gmail',
-      auth: {
-        user: process.env.GMAIL_USER,
-        pass: process.env.GMAIL_PASSWORD,
-      },
-    })
-
-    // Compose the email content
-    const mailOptions = {
-      from: process.env.GMAIL_USER,
-      to: user.email,
-      subject: 'Cash Out Ready',
-      text: `Hi, ${user.firstName} ${user.lastName} \n\nCongratuations! Your cashout is ready \n\n  Please navigate to your account and cashout your funds from your group and get ready for the next contribution if your group will still be active. Thank you and enjoy your funds.`,
-    }
-
-    // Send the email
-    transporter.sendMail(
-      mailOptions,
-      (err: Error | null, info: { response: string }) => {
-        if (err) {
-          console.log(
-            `Failed to send cashout email. Please try again later.`,
-            err
-          )
-        } else {
-          console.log(
-            'Cashout email has been sent to your email if you have an account with us: ' +
-              info.response
-          )
-        }
-      }
-    )
-  } else {
-    console.log('No next beneficiary to notify.')
-  }
-}
-
-async function triggerNextBeneficiary(userId: string, user: User) {
-  console.log('Triggering next beneficiary...')
-  const nextBeneficiary = await determineNextBeneficiary(userId)
-  await notifyBeneficiary(nextBeneficiary, user)
-}
-
-async function createNextCycle(newGroup: { id: string }) {
-  try {
-    console.log('Creating next cycle...')
-    // Retrieve the current group
-    const currentGroup = await Groups.findByPk(newGroup.id)
-    if (!currentGroup) {
-      throw new Error('Group not found')
-    }
-
-    // Reset contribution counters
-    currentGroup.amountContributedWithinFrequency = 0
-    currentGroup.totalAmountWithdrawn = 0
-
-    // Save the updated group
-    await currentGroup.save()
-
-    console.log('Next cycle created successfully.')
-  } catch (error) {
-    console.error('Error creating next cycle:', error)
-    // return HTTP_STATUS_HELPER[HTTP_STATUS_CODE.INTERNAL_SERVER](res, {
-    //   message: 'Error creating next cycle',
-    //   error: error,
-    // })
-  }
-}
-
-async function determineNextBeneficiary(
-  userId: string
-): Promise<string | null> {
-  const nextMember = await GroupMembers.findOne({
-    where: {
-      userId,
-    },
-    order: [['sequence', 'ASC']],
-  })
-
-  return nextMember ? nextMember.userId : null
-}
-
-function calculateFrequencyDuration(
-  maxParticipants: number,
-  frequency: 'daily' | 'weekly' | 'monthly'
-): number {
-  // Convert frequency to milliseconds
-  let frequencyInMilliseconds
-  switch (frequency) {
-    case 'daily':
-      frequencyInMilliseconds = 24 * 60 * 60 * 1000
-      break
-    case 'weekly':
-      frequencyInMilliseconds = 7 * 24 * 60 * 60 * 1000
-      break
-    case 'monthly':
-      frequencyInMilliseconds = 28 * 24 * 60 * 60 * 1000
-      break
-    default:
-      throw new Error('Invalid frequency')
-  }
-
-  // Calculate cycle duration based on the number of participants
-  const cycleDuration = maxParticipants * frequencyInMilliseconds
-
-  return cycleDuration
-}
+import { Duration, add, format } from 'date-fns'
+import { triggerNextBeneficiary } from './helpers'
+import { createNextCycle } from './helpers'
+import cron from 'node-cron'
+import { getCronExpressionFromFrequency } from './helpers'
 
 export const createGroup = async (req: RequestExt, res: Response) => {
   const { _user: user, _userId: userId, ...rest } = req.body
@@ -189,31 +74,43 @@ export const createGroup = async (req: RequestExt, res: Response) => {
       automaticRestartCycle: _data.automaticRestartCycle,
     }
 
-    const frequencyDuration = calculateFrequencyDuration(
-      _data.maxNumberOfParticipants,
-      _data.frequency
-    )
+    const currentDate = new Date()
+    let nextCashoutDate: Date
 
-    const formattedDate = format(
-      addMilliseconds(new Date(), frequencyDuration),
-      "yyyy-MM-dd'T'HH:mm:ss"
-    )
+    switch (_data.frequency) {
+      case 'daily':
+        nextCashoutDate = add(currentDate, { days: 1 } as Duration)
+        break
+      case 'weekly':
+        nextCashoutDate = add(currentDate, { weeks: 1 } as Duration)
+        break
+      case 'monthly':
+        nextCashoutDate = add(currentDate, { months: 1 } as Duration)
+        break
+      default:
+        throw new Error('Invalid frequency')
+    }
+
+    const formattedDate = format(nextCashoutDate, 'yyyy-MM-dd')
 
     const automaticRestartCycleEnabled = newGroup.automaticRestartCycle || false
 
-    setTimeout(async () => {
+    const cronExpression = getCronExpressionFromFrequency(
+      _data.frequency,
+      new Date()
+    )
+    const cronDate = cron.schedule(cronExpression, async () => {
       try {
         await triggerNextBeneficiary(userId, user)
         if (automaticRestartCycleEnabled) {
           await createNextCycle(newGroup)
         }
       } catch (error) {
-        return HTTP_STATUS_HELPER[HTTP_STATUS_CODE.INTERNAL_SERVER](res, {
-          message: 'Error during contribution cycle:',
-          error,
-        })
+        console.error('Error during contribution cycle:', error)
       }
-    }, frequencyDuration)
+    })
+
+    cronDate.start()
 
     const createdGroup = await Groups.create(newGroup)
